refactoring apiIntegrationTestUtil.ts
